Move the Pointer class to the ffi package
