+++ src/main/java/org/truffleruby/platform/NativePlatform.java
@@ -9,6 +9,7 @@
  */
 package org.truffleruby.platform;
 
+import org.truffleruby.extra.ffi.Pointer;
 import org.truffleruby.platform.posix.ClockGetTime;
 import org.truffleruby.platform.posix.Sockets;
 import org.truffleruby.platform.posix.Threads;
